+++ src/components/BaseStyles/typography.js
@@ -56,4 +56,32 @@ const SectionMessage = styled(motion.p)`
   `}
 `
 
-export { SectionTitle, SectionSubTitle, SectionMessage }
+const JpnName = styled(motion.span)`
+  position: absolute;
+  top: 0;
+  right: 0;
+  z-index: -1;
+  // text
+  word-break: break-all;
+  line-height: 1;
+  text-transform: uppercase;
+  text-align: center;
+  font-size: 3rem;
+  font-weight: bold;
+  user-select: none;
+  width: 1em;
+
+  ${({ theme }) => css`
+    color: ${theme.jpnName.color};
+
+    @media ${theme.device.xxs} {
+      display: none;
+    }
+    color: black;
+    @media ${theme.device.md} {
+      display: inline-block;
+    }
+  `}
+`
+
+export { SectionTitle, SectionSubTitle, SectionMessage, JpnName }
